Made-with: Cursor

+++ cs3305-team6/src/main/java/com/team6/pojo/Product.java
@@ -1,5 +1,6 @@
 package com.team6.pojo;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -13,18 +14,18 @@ public class Product {
     private String name;
     private String brand;
     private String imageUrl;
-    /** Price in smallest unit (e.g. cents). 12.99 EUR → 1299 */
-    private Long price;
+    /** Current price (DECIMAL, e.g. 12.99) */
+    private BigDecimal price;
     private String currency;
-    /** Per 100g, scaled by 100. 5.25g → 525 */
-    private Long energyKcal;
-    private Long fat;
-    private Long saturatedFat;
-    private Long carbohydrates;
-    private Long sugars;
-    private Long fiber;
-    private Long proteins;
-    private Long salt;
+    /** Nutrition per 100g (DECIMAL) */
+    private BigDecimal energyKcal;
+    private BigDecimal fat;
+    private BigDecimal saturatedFat;
+    private BigDecimal carbohydrates;
+    private BigDecimal sugars;
+    private BigDecimal fiber;
+    private BigDecimal proteins;
+    private BigDecimal salt;
     private String nutriScore;
     private String source;
     private String sourceUrl;
@@ -73,11 +74,11 @@ public class Product {
         this.imageUrl = imageUrl;
     }
 
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
@@ -89,67 +90,67 @@ public class Product {
         this.currency = currency;
     }
 
-    public Long getEnergyKcal() {
+    public BigDecimal getEnergyKcal() {
         return energyKcal;
     }
 
-    public void setEnergyKcal(Long energyKcal) {
+    public void setEnergyKcal(BigDecimal energyKcal) {
         this.energyKcal = energyKcal;
     }
 
-    public Long getFat() {
+    public BigDecimal getFat() {
         return fat;
     }
 
-    public void setFat(Long fat) {
+    public void setFat(BigDecimal fat) {
         this.fat = fat;
     }
 
-    public Long getSaturatedFat() {
+    public BigDecimal getSaturatedFat() {
         return saturatedFat;
     }
 
-    public void setSaturatedFat(Long saturatedFat) {
+    public void setSaturatedFat(BigDecimal saturatedFat) {
         this.saturatedFat = saturatedFat;
     }
 
-    public Long getCarbohydrates() {
+    public BigDecimal getCarbohydrates() {
         return carbohydrates;
     }
 
-    public void setCarbohydrates(Long carbohydrates) {
+    public void setCarbohydrates(BigDecimal carbohydrates) {
         this.carbohydrates = carbohydrates;
     }
 
-    public Long getSugars() {
+    public BigDecimal getSugars() {
         return sugars;
     }
 
-    public void setSugars(Long sugars) {
+    public void setSugars(BigDecimal sugars) {
         this.sugars = sugars;
     }
 
-    public Long getFiber() {
+    public BigDecimal getFiber() {
         return fiber;
     }
 
-    public void setFiber(Long fiber) {
+    public void setFiber(BigDecimal fiber) {
         this.fiber = fiber;
     }
 
-    public Long getProteins() {
+    public BigDecimal getProteins() {
         return proteins;
     }
 
-    public void setProteins(Long proteins) {
+    public void setProteins(BigDecimal proteins) {
         this.proteins = proteins;
     }
 
-    public Long getSalt() {
+    public BigDecimal getSalt() {
         return salt;
     }
 
-    public void setSalt(Long salt) {
+    public void setSalt(BigDecimal salt) {
         this.salt = salt;
     }
